+++ apps/ledger-live-desktop/src/newArch/Collectibles/Nfts/Gallery/index.tsx
@@ -5,7 +5,7 @@ import Box from "~/renderer/components/Box";
 import Text from "~/renderer/components/Text";
 import Button from "~/renderer/components/Button";
 import { Flex, Icons } from "@ledgerhq/react-ui";
-import { TableLayout } from "../../components";
+import { TableLayout } from "LLD/Collectibles/components";
 import { useTranslation } from "react-i18next";
 import { Account } from "@ledgerhq/types-live";
 import Spinner from "~/renderer/components/Spinner";
