+++ apps/ledger-live-desktop/src/newArch/Collectibles/components/Collection/TableHeader.tsx
@@ -2,7 +2,7 @@ import { Flex } from "@ledgerhq/react-ui";
 import React from "react";
 import { useTranslation } from "react-i18next";
 import { TableHeader as HeaderContainer } from "~/renderer/components/TableContainer";
-import { TableHeaderProps as Props } from "../../types/Collection";
+import { TableHeaderProps as Props } from "LLD/Collectibles/types/Collection";
 
 const TableHeader: React.FC<Props> = ({ titleKey, actions }) => {
   const { t } = useTranslation();
