+++ apps/ledger-live-desktop/src/newArch/Collectibles/components/Gallery/TokenList/index.tsx
@@ -5,11 +5,11 @@ import { useSelector } from "react-redux";
 import { collectiblesViewModeSelector } from "~/renderer/reducers/settings";
 import Item from "./Item";
 import { Layout, LayoutKey } from "LLD/Collectibles/types/Layouts";
-import { TokenListProps } from "~/newArch/Collectibles/Nfts/components/TokensList/useTokensListModel";
+import { TokensListProps } from "LLD/Collectibles/types/TokensList";
 
 type Props = {
-  account: TokenListProps["account"];
-  collectibles: TokenListProps["formattedNfts"];
+  account: TokensListProps["account"];
+  collectibles: TokensListProps["formattedNfts"];
   onHideCollection?: () => void;
   onItemClick: (id: string) => void;
 };
