+++ apps/ledger-live-desktop/src/newArch/Collectibles/hooks/useNftLinks.tsx
@@ -15,7 +15,7 @@ import CustomImage from "~/renderer/screens/customImage";
 import { ContextMenuItemType } from "~/renderer/components/ContextMenu/ContextMenuWrapper";
 import { devicesModelListSelector } from "~/renderer/reducers/settings";
 import { safeList } from "LLD/Collectibles/utils/useSafeList";
-import { ItemType } from "LLD/Collectibles/types/Links";
+import { ItemType } from "LLD/Collectibles/types/enum/Links";
 
 const linksPerCurrency: Record<
   string,
