+++ apps/ledger-live-desktop/src/newArch/Collectibles/types/enum/Collectibles.ts
@@ -1,0 +1,4 @@
+export enum CollectibleTypeEnum {
+  NFT = "NFT",
+  Ordinal = "Ordinal",
+}
