+++ apps/ledger-live-desktop/src/newArch/Collectibles/types/enum/DetailDrawer.ts
@@ -1,0 +1,10 @@
+export enum FieldStatus {
+  Loading = "loading",
+  Loaded = "loaded",
+  Error = "error",
+  NoData = "nodata",
+}
+export enum ItemType {
+  Separator = "separator",
+  ExternalLink = "external",
+}
