+++ apps/ledger-live-desktop/src/newArch/Collectibles/types/enum/Links.ts
